self/requirement_home.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==10.4.0
PyJWT==2.9.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
sqlparse==0.5.1
typing==3.7.4.3
urllib3==2.2.3
